Refactor transaction logic and enhance account validations
Updated transaction methods to handle `double` instead of `int` for more precise operations. Improved error handling in deposit, withdraw, and transfer processes, adding specific exceptions for invalid accounts, insufficient funds, and transaction limits. Adjusted GUI to provide clearer feedback messages and fixed balance display updates.

+++ src/models/BankAccount.java
@@ -79,7 +79,7 @@ public class BankAccount {
   }
 
   // Deposit method
-  public void deposit(int amount) throws GlobalExceptionHandler {
+  public void deposit(double amount) throws GlobalExceptionHandler {
     if (amount <= 0) {
       throw new InvalidAmountException("Amount must be positive");
     }
@@ -89,7 +89,7 @@ public class BankAccount {
   }
 
   // Withdraw method
-  public void withdraw(int amount) throws GlobalExceptionHandler {
+  public void withdraw(double amount) throws GlobalExceptionHandler {
     if (amount <= 0) {
       throw new InvalidAmountException("Amount must be positive");
     }
